Added few exeptions.

+++ src/main/java/com/altindag/hobbyproject/service/PersonService.java
@@ -2,16 +2,15 @@ package com.altindag.hobbyproject.service;
 
 import com.altindag.hobbyproject.domain.Person;
 import com.altindag.hobbyproject.repository.PersonRepository;
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PersonService {
 
-    @Getter
     private List<Person> people;
 
     @Autowired
@@ -21,17 +20,27 @@ public class PersonService {
         this.people = people;
     }
 
+    public List<Person> getPeople() {
+        return personRepository.findAll();
+    }
+
     public void addPerson(Person person) {
+        Optional<Person> idOptional = personRepository.findById(person.getId());
+        if (idOptional.isPresent()){
+            throw new IllegalStateException("id taken");
+        }
         personRepository.save(person);
     }
 
     public void deletePerson(Long id){
+        boolean present = personRepository.existsById(id);
+        if(!present){
+            throw new IllegalStateException(String.format("Person with id: %s does not exists", id));
+        }
         personRepository.deleteById(id);
     }
 
     public void updatePerson(Person person){
-
-        personRepository.findById(person.getId());
         int idx = 0;
         Long id = 0L;
